Working /etc config
This is messy, because it means the config construction script has to
have root access to create the config dir and set permissions.
Additionally, the intention with systemd seems to be that
ConfigurationDirectory is world-readable, which is not suitable for us
because it contains secrets.

Look into LoadCredential for alternatives.